pyproject.toml
[project]
name = "acs-sdk"
version = "0.3.0"
description = "Python SDK for Accelerated Cloud Storage service"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "AcceleratedCloudStorage", email = "sales@acceleratedcloudstorage.com" }
]
urls = { "Homepage" = "https://github.com/AcceleratedCloudStorage/acs-sdk-python" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "grpcio>=1.70.0",
    "grpcio-tools>=1.70.0",
    "PyYAML>=6.0.2",
    "protobuf>=5.29.3",
    "iniconfig>=2.0.0",
    "pluggy>=1.5.0",
    "protobuf>=5.29.3",
    "setuptools>=75.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=22.0",
    "mypy>=0.900",
    "grpc-stubs>=1.53.0",
]

[tool.setuptools]
packages = [
    "acs_sdk",
    "acs_sdk.client",
    "acs_sdk.internal",
    "acs_sdk.internal.generated"
]

[tool.setuptools.package-data]
"acs_sdk" = ["internal/ca-chain.pem"]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow",
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv"
]

# Formatting settings (same as Black)
line-length = 79
indent-width = 4
target-version = "py38"

[tool.ruff.lint]
# Enable only selected Pyflakes (F) and pycodestyle error codes (E4, E7, E9)
select = ["E4", "E7", "E9", "F", "I", "UP"]
ignore = []
fixable = ["ALL"]
unfixable = []
# Allow unused variables when names start with one or more underscores.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Use double quotes, spaces for indents and preserve trailing commas
quote-style = "preserve"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"